refactor(wallet-filter-sidebar): migrate component to Chakra
